% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_sst_change.R
\name{get_all_rasters}
\alias{get_all_rasters}
\title{Calculate a raster brick object containing the average sst, decadal rate of
linear temperature change, spatial gradient, and velocity of climate change
over a time period specified by years.}
\usage{
get_all_rasters(hadsst_raster, years = 1969:2009)
}
\arguments{
\item{hadsst_raster}{the HadISST data as a raster object.}

\item{years}{a numeric vector specifying the years over which metrics of
sea surface temperature will be calculated.}
}
\value{
a multi-layer raster brick containing the average sea surface
temperature, decadal rate of linear temperature change, magnitude of the
spatial gradient, magnitude of the velocity of climate change, and the
angle of the spatial gradient and velocity of climate change (same angle).
}
\description{
Calculate a raster brick object containing the average sst, decadal rate of
linear temperature change, spatial gradient, and velocity of climate change
over a time period specified by years.
}

