% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_ssts.R
\name{get_average_sst}
\alias{get_average_sst}
\title{Calculate the average temperature over a range of years}
\usage{
get_average_sst(hadsst_raster, years = 1969:2011)
}
\arguments{
\item{hadsst_raster}{the HadISST data as a raster object.}

\item{years}{a numeric vector specifying the years used to calculate the
sea surface temperature.}
}
\value{
a single-layer raster object with the average temperature for each cell of
the time period specified in years.
}
\description{
Calculate the average temperature over a range of years
}

