% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_sst_change.R
\name{get_NS_diffs}
\alias{get_NS_diffs}
\title{Calculate the north to south spatial gradient in temperature between
neighbouring cells.}
\usage{
get_NS_diffs(avg_raster)
}
\arguments{
\item{avg_raster}{is a single-layer raster containing the average sea surface
temperature over a range of years.}
}
\value{
a single-layer raster with the north-south spatial gradient and
corrected for distance by latitude.
}
\description{
Calculate the north to south spatial gradient in temperature between
neighbouring cells.
}
\seealso{
\code{\link{get_average_sst}}
}
