% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_sst_change.R
\name{get_spatial_gradient}
\alias{get_spatial_gradient}
\title{Calculate the spatial gradient of temperature change.}
\usage{
get_spatial_gradient(NS_gradient, WE_gradient)
}
\arguments{
\item{NS_gradient}{single-layer raster containing the north to south
spatial gradient between neighbouring cells}

\item{WE_gradient}{single-layer raster containing the west to east
spatial gradient between neighbouring cells}
}
\value{
a two-layer raster containing 'spatial_gradient', the magnitude of the spatial gradient and 'angle', the direction of the spatial gradient.
}
\description{
Calculate the spatial gradient of temperature change.
}
\references{
Loarie, S. R., P. B. Duffy, H. Hamilton, G. P. Asner, C. B. Field, and D. D. Ackerly. 2009. The velocity of climate change. Nature 462:1052–5.
Burrows, M. T., D. S. Schoeman, L. B. Buckley, P. Moore, E. S. Poloczanska, K. M. Brander, C. Brown, J. F. Bruno, C. M. Duarte, B. S. Halpern, J. Holding, C. V Kappel, W. Kiessling, M. I. O’Connor, J. M. Pandolfi, C. Parmesan, F. B. Schwing, W. J. Sydeman, and A. J. Richardson. 2011. The pace of shifting climate in marine and terrestrial ecosystems. Science (New York, N.Y.) 334:652–5.
}
\seealso{
\code{hadsstr::get_average_sst}, \code{hadsstr::get_vocc}
}

