% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_sst_change.R
\name{get_sst_linear_change}
\alias{get_sst_linear_change}
\title{Calculate the decadal rate of temperature change over a given time period.}
\usage{
get_sst_linear_change(hadsst_raster = b, years = 2000:2011)
}
\arguments{
\item{hadsst_raster}{the HadISST data as a raster object.}

\item{years}{a numeric vector specifying the year(s) over which the decadal rate of
linear temperature change will be calculated.}
}
\value{
a single layer raster with the decadal rate of linear temperature cell
for each cell of the time period specified in years
}
\description{
Calculate the decadal rate of temperature change over a given time period.
}

