% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_sst_change.R
\name{get_vocc}
\alias{get_vocc}
\title{Calculate the velocity of climate change over a given set of years}
\usage{
get_vocc(hadsst_raster, years = 1969:2009, truncate = FALSE)
}
\arguments{
\item{hadsst_raster}{the HadISST data as a raster object.}

\item{years}{a numeric vector specifying the years over which the velocity of
climate change will be calculated.}

\item{truncate}{logical expression indicating whether minimum and maximum
values should be constrained to -200 < x < 200. Default is TRUE.}
}
\value{
a two-layer raster containing 'velocity_mag' and 'velocity_angle',
the magnitude and direction of the velocity of climate change, respectively.
}
\description{
Calculate the velocity of climate change over a given set of years
}
\references{
Loarie, S. R., P. B. Duffy, H. Hamilton, G. P. Asner, C. B. Field, and D. D. Ackerly. 2009. The velocity of climate change. Nature 462:1052–5.
Burrows, M. T., D. S. Schoeman, L. B. Buckley, P. Moore, E. S. Poloczanska, K. M. Brander, C. Brown, J. F. Bruno, C. M. Duarte, B. S. Halpern, J. Holding, C. V Kappel, W. Kiessling, M. I. O’Connor, J. M. Pandolfi, C. Parmesan, F. B. Schwing, W. J. Sydeman, and A. J. Richardson. 2011. The pace of shifting climate in marine and terrestrial ecosystems. Science (New York, N.Y.) 334:652–5.
}
\seealso{
\code{hadsstr::get_average_sst}
}
