% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_sst_change.R
\name{get_WE_diffs}
\alias{get_WE_diffs}
\title{Calculate the west to east spatial gradient in temperature between neighbouring
cells (corrected for distance by latitude).}
\usage{
get_WE_diffs(avg_raster)
}
\arguments{
\item{avg_raster}{is a single-layer raster containing the average sea surface
temperature over a range of years.}
}
\value{
a single-layer raster with the west-east spatial gradient and
corrected for distance by latitude.
}
\description{
Calculate the west to east spatial gradient in temperature between neighbouring
cells (corrected for distance by latitude).
}
\seealso{
\code{hadsstr::get_average_sst}
}
