% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getClimateLatLon}
\alias{getClimateLatLon}
\title{getClimateChange}
\usage{
getClimateLatLon(sstMatsObj, lat, lon, measure = "velocity")
}
\arguments{
\item{sstMatsObj}{An object containing latitude, longitude, and matrices of climate information from \code{climateChangeMats}}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{measure}{A string defining which type of information. Defaults to "velocity". Can be "average", "spatialGrad",
"angle", "velocity", "NS", "EW".}
}
\value{
Returns a single value
}
\description{
\code{getClimateLatLon} Get information about climate change at a particular latitude and longitude
}
\details{
Requires that you have generated climate change matrices
}
\examples{
sstData <- loadHadSST(directory="./", hadsstFilename="HadISST_sst.nc")

climateChangeMats <- getClimateChange(sstData)

getClimateLatLon(climateChangeMats, -50.232, -100.55, "linearChange")
}
\author{
Jarrett Byrnes.
}

