% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSSTAnnualArray}
\alias{getSSTAnnualArray}
\title{getSSTAnnualArray}
\usage{
getSSTAnnualArray(sstObj, years = 1969:2009)
}
\arguments{
\item{sstObj}{The HadSST Data object}

\item{years}{A range of years to use to examine climate change}
}
\value{
Returns an array with each slice containing one matrix of temperatures
averaged over a year
}
\description{
\code{getSSTAnnualArray} Collapse monthly HadSST data into annual averages
}
\details{
Requires that you have loaded a HADSST data file.
}
\examples{
sstData <- loadHadSST(directory="./", hadsstFilename="HadISST_sst.nc")
summary(sstData)

getSSTAnnualArray(sstData, years=1984:2004)
}
\author{
Jarrett Byrnes.
}

