% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSSTAvgMat}
\alias{getSSTAvgMat}
\title{getSSTAvgMat}
\usage{
getSSTAvgMat(sstObj, years = 1969:2009)
}
\arguments{
\item{years}{A range of years to use to examine climate change}

\item{sstAnnualArray}{An array of SST values at different points in time}
}
\value{
Returns a matrix of average temperature values
}
\description{
\code{getSSTAvgMat} Gets the matrix describing average temperature across years
}
\details{
Requires that you have loaded a HADSST data file. Will return a matrix of average temperature
of all lat/longs in all selected years
}
\examples{
sstData <- loadHadSST(directory="./", hadsstFilename="HadISST_sst.nc")

getSSTAvgMat(sstData, years=1969:2009)
}
\author{
Jarrett Byrnes.
}
\seealso{
\code{\link{getSSTLinearChangeMatFromArray}}, \code{\link{getSSTAvgMatFromArray}}
}

