% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSSTAvgMatFromArray}
\alias{getSSTAvgMatFromArray}
\title{getSSTAvgMatFromArray}
\usage{
getSSTAvgMatFromArray(sstAnnualArray)
}
\arguments{
\item{sstAnnualArray}{An array of SST values at different points in time}
}
\value{
Returns a matrix of average temperature values
}
\description{
\code{getSSTAvgMatFromArray} Gets the matrix describing average temperature across years
from an array of geospatial temperature values
}
\details{
Requires that you have loaded a HADSST data file. Will return a matrix of average temperature
of all values in the third dimension of the array
}
\examples{
sstData <- loadHadSST(directory="./", hadsstFilename="HadISST_sst.nc")

getSSTAvgMatFromArray(sstData$sstArray)
}
\author{
Jarrett Byrnes.
}

