% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSSTLinearChangeMatFromArray}
\alias{getSSTLinearChangeMatFromArray}
\title{getSSTLinearChangeMatFromArray}
\usage{
getSSTLinearChangeMatFromArray(sstAnnualArray, years = 1969:2009)
}
\arguments{
\item{sstAnnualArray}{The HadSST Data object}

\item{years}{The range of time in years for the array being used.}
}
\value{
Returns a matrix of average temperature values
}
\description{
\code{getSSTLinearChangeMatFromArray} Gets the matrix with linear change coefficients for
temperature in each grid cell
}
\details{
Requires that you have loaded a HADSST data file. Will return a matrix of decadal slopes
for all lat/longs in all selected years. Note, for regressions below the annual level, the 'years'
argument must be in decimal years.
}
\examples{
sstData <- loadHadSST(directory="./", hadsstFilename="HadISST_sst.nc")

years <- 1960:2009

sstAnnualArray <- getSSTAnnualArray(sstData, years)

getSSTLinearChangeMatFromArray(sstData$sstAnnualArray, years)
}
\author{
Jarrett Byrnes.
}
\seealso{
\code{\link{getClimateChange}}, \code{\link{getSSTAnnualArray}}
}

