% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hadsst_data.R
\name{load_hadsst}
\alias{load_hadsst}
\title{Load the netcdf sea surface temperature data.}
\usage{
load_hadsst(file = "./HadISST_sst.nc")
}
\arguments{
\item{path}{to HadISST data in the netcdf format}
}
\description{
Load the netcdf sea surface temperature data.
}

