% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadHadSST}
\alias{loadHadSST}
\title{loadHadSST}
\usage{
loadHadSST(directory = "./", hadsstFilename = "HadISST_sst.nc",
  latFlip = TRUE)
}
\arguments{
\item{directory}{Here is the SST netcdf file found?}

\item{hadsstFilename}{The name of the netcdf file}

\item{latFlip}{Some of the HADSST data files have the sign of the latitude flipped.
Defaults to \code{TRUE}, and flips the array by latitude}
}
\value{
Returns a hadsstData object containing an array of measurements, the latitudes
and longitudes of data as well as the dates

\code{sstArray} The array of SST by longitude, latitude, and time

\code{lon} The values of longitude

\code{lat} The values of latitude

\code{tdates} The dates of each slice of the data
}
\description{
\code{loadHadSST} Loads data from a HADSST netcdf file and returns a hadsstDATA object
}
\details{
Requires that you have downloaded and unpacked a HADSST data file.
}
\examples{
sstData <- loadHadSST(directory="./", hadsstFilename="HadISST_sst.nc")
summary(sstData)
}
\author{
Jarrett Byrnes.
}

