% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_sst_change.R
\name{select_raster}
\alias{select_raster}
\title{Select layers from the raster brick object created using \code{hadsstr::get_all_rasters}.
Layers contain the average sst, decadal rate of linear temperature change,
spatial gradient, and velocity of climate change for a set of years.}
\usage{
select_raster(all_rasters, layer_name)
}
\arguments{
\item{layer_name}{character vector containing the name(s) of the layers}

\item{raster}{brick created using the \code{hadsstr::get_all_rasters} the HadISST data as a raster object.}
}
\value{
a raster or raster brick containing layers selected in 'layer_name'.
}
\description{
Select layers from the raster brick object created using \code{hadsstr::get_all_rasters}.
Layers contain the average sst, decadal rate of linear temperature change,
spatial gradient, and velocity of climate change for a set of years.
}
