% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_raster.R
\name{subset_years}
\alias{subset_years}
\title{Subset the hadsst raster by year.}
\usage{
subset_years(hadsst_raster, years = 2000:2010, brick = TRUE)
}
\arguments{
\item{hadsst_raster}{the HadISST data as a raster object.}

\item{years}{a numeric vector specifying the year(s) over which the decadal rate of
linear temperature change will be calculated.}

\item{brick}{a boolean field specifying if you would like to return the subsetted
raster as a list of rasters or as a single brick of rasters.}
}
\value{
a single layer raster with the decadal rate of linear temperature cell
for each cell of the time period specified in years
}
\description{
Subset the hadsst raster by year.
}
