% Generated by roxygen2 (4.0.2): do not edit by hand
\name{truncateMat}
\alias{truncateMat}
\title{truncateMat}
\usage{
truncateMat(mat, upper, lower)
}
\arguments{
\item{mat}{The matrix to be turncated}

\item{upper}{The upper limit}

\item{lower}{lower limit}
}
\value{
Returns a \code{matrix}
}
\description{
\code{truncateMat} Convenience function to truncate extreme values in a matrix
}
\details{
This is ideal for plotting matrices with extreme high and low values
}
\examples{
mat <- matrix(1:100, nrow=10)
truncateMat(mat, 80, 20)
}
\author{
Jarrett Byrnes.
}

